LayoutTests/imported/w3c/web-platform-tests/credential-management/idl.https-expected.txt


FAIL CredentialsContainer interface: existence and properties of interface object assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface object length assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface object name assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface: operation get(CredentialRequestOptions) assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface: operation store(Credential) assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface: operation create(CredentialCreationOptions) assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer interface: operation preventSilentAccess() assert_own_property: self does not have own property "CredentialsContainer" expected property "CredentialsContainer" missing
FAIL CredentialsContainer must be primary interface of navigator.credentials assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of navigator.credentials assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CredentialsContainer interface: navigator.credentials must inherit property "get(CredentialRequestOptions)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CredentialsContainer interface: calling get(CredentialRequestOptions) on navigator.credentials with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CredentialsContainer interface: navigator.credentials must inherit property "store(Credential)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CredentialsContainer interface: calling store(Credential) on navigator.credentials with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CredentialsContainer interface: navigator.credentials must inherit property "create(CredentialCreationOptions)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CredentialsContainer interface: calling create(CredentialCreationOptions) on navigator.credentials with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL CredentialsContainer interface: navigator.credentials must inherit property "preventSilentAccess()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PasswordCredential interface: existence and properties of interface object assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface object length assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface object name assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: existence and properties of interface prototype object assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential interface: attribute password assert_own_property: self does not have own property "PasswordCredential" expected property "PasswordCredential" missing
FAIL PasswordCredential must be primary interface of new PasswordCredential({ id: "id", password: "pencil", iconURL: "https://example.com/", name: "name" }) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: PasswordCredential"
FAIL Stringification of new PasswordCredential({ id: "id", password: "pencil", iconURL: "https://example.com/", name: "name" }) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: PasswordCredential"
FAIL PasswordCredential interface: new PasswordCredential({ id: "id", password: "pencil", iconURL: "https://example.com/", name: "name" }) must inherit property "password" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: PasswordCredential"
FAIL FederatedCredential interface: existence and properties of interface object assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface object length assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface object name assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: existence and properties of interface prototype object assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: attribute provider assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential interface: attribute protocol assert_own_property: self does not have own property "FederatedCredential" expected property "FederatedCredential" missing
FAIL FederatedCredential must be primary interface of new FederatedCredential({ id: "id", provider: "https://example.com", iconURL: "https://example.com/", name: "name" }) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: FederatedCredential"
FAIL Stringification of new FederatedCredential({ id: "id", provider: "https://example.com", iconURL: "https://example.com/", name: "name" }) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: FederatedCredential"
FAIL FederatedCredential interface: new FederatedCredential({ id: "id", provider: "https://example.com", iconURL: "https://example.com/", name: "name" }) must inherit property "provider" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: FederatedCredential"
FAIL FederatedCredential interface: new FederatedCredential({ id: "id", provider: "https://example.com", iconURL: "https://example.com/", name: "name" }) must inherit property "protocol" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: FederatedCredential"
